Use Property.isDefined() to remove the need for HasProperty.
In the Kernel API, there is no need for nodeHasProperty or
relationshipHasProperty, since they are implemented as *GetProperty()
followed by isDefined() on the returned Property (or NoProperty)
object anyhow. Since this is an internal API, there is more value in
keeping it small, so we might as well do that in the places of use
instead.