BatchInserter support for LabelScanStore
on a clean shutdown the batch inserter populates the selected label scan
store. It's doing so in the same store scan as the one populating all
schema indexes. For that NeoStoreIndexStoreView changed to support one
property update visitor and one label update visitor and invites them both
for a single store scan.

Introduced a NeoStoreProvider which take on the same role in the batch
inserter case as XaDataSourceManager does in the GDS case. A label scan
store may need to access the NeoStore as a dependency and the actual
NeoStore instance isn't guaranteed to exist at startup or when the label
scan store (via an extension) is initialized.