Fixes a FullTxData bug analyzing a query
The FullTxData instance storing state about removed entities where the
removals where made w/o value and potentially also w/o key would try to
decorate incoming queries with additional clauses including those entities
so that they will be filtered out. In doing that it looks at the query to
find out which keys it queries. There was a case where an incoming
query "*:*", which doesn't have a particular key and trying to find it
resulted in exception.

Solution is to add a specific condition for this query type so that no
special handling is done, since that query will match all entities either
way.