Adds the ability to do label scan search as a Term Query.
Potentially this is more efficient than using Range Queries to emulate "key presence" queries.

This is done by having a separate field that is indexed only (not stored in the document) for searches,
and then not indexing the bitmap field. So storage wise, this implementation should be about the same.
Although it is possible that both in terms of storage overhead, and query efficiency having multiple
different indexed fields is sub optimal, in which case this implementation would be better.
This is to be determined by benchmarks.