Don't activate indexes during recovery.
Fix for a failure in recovery of Unique Indexes where recovery would
crash with the message "cannot await recovering index" when recovering a
store where the active log both creates and drops the index in question.

The correctness of this is based on the assumption that there are other
processes (already in place) to make sure that the indexes are activated
post-recovery.