Refactoring of Predicate.rewrite
This method didn't follow the signature from AstNode - instead of returning an Expression,
rewrite returned a Predicate. This meant that it is impossible to rewrite predicates that
are used as expressions to something other than Predicate.

When Cypher is caching expressions, it uses this rewriting. This is done when using
complicated expressions with sub-expressions that are aggregate expressions mixed with
sub-expressions that are not aggregating. In these cases, the rewriting would fail and
manifested itself through a "Missing identifier" error.

 The solution is to move the expectation that rewrite will return a Predicate to the user
 of the method.