Use TargetDirectory to run each test case in its own directory.
This test involves servers failing to startup due to timeout.
In some situations the server will fail to shutdown fully in these
situations, leaving file locks that block the next test case.

A full fix for this issue would involve changing the lifecycle of
NeoStoreXaDataSource so that it's subordinate components are managed
through LifeSupport, which would give better assurance of shutdown
on failure to start. However, a change of that cope is too large for a
maintenance release, and this is only really an issue in tests where we
run multiple failed servers in the same JVM.