Uses a DEV_NULL StringLogger instead of SYSTEM
due to, most likely, a problem with the test or even SubProcess that the
piping of System.out/System.err from the sub process where those buffers
fill up in the sub process since the parent process doesn't pull that
data. When those buffers have filled up, then any write to its System.xxx
will block awaiting the buffer to drain, which it doesn't.

The current fix is just to not use System.out in this test, however a
future fix should be to look at and fix the issue in SubProcess itself.