Remove a catch block that was suppressing failures in a test and make the test more robust.
The catch block in HTTPLoggingDocIT was suppressing all failures. Once
it had been removed it became apparent that the test was
flaky. Specifically it almost always failed in my IDE when run
alone (although it passed when run with the other tests in the file). I
believe the flakiness was because the log file wasn't always flushed to
disk before the assertion was reached.